+++ src/future/backports/__init__.py
@@ -6,3 +6,5 @@ from future.standard_library import import_top_level_modules
 
 if sys.version_info[0] == 3:
     import_top_level_modules()
+
+from .misc import ceil, OrderedDict, Counter, check_output, count