+++ src/future/types/newrange.py
@@ -21,6 +21,10 @@ From Dan Crosta's README:
 from collections import Sequence, Iterator
 from itertools import islice
 
+from future.backports.misc import count   # with step parameter on Py2.6
+# For backward compatibility with python-future versions < 0.14.4:
+from future.backports.misc import _count
+
 
 class newrange(Sequence):
     """
@@ -142,20 +146,13 @@ class range_iterator(Iterator):
     """An iterator for a :class:`range`.
     """
     def __init__(self, range_):
-        self._stepper = islice(_count(range_.start, range_.step), len(range_))
+        self._stepper = islice(count(range_.start, range_.step), len(range_))
 
     def __iter__(self):
         return self
 
     def next(self):
         return next(self._stepper)
-
-
-# itertools.count in Py 2.6 doesn't accept a step parameter
-def _count(start=0, step=1):
-    while True:
-        yield start
-        start += step
 
 
 __all__ = ['newrange']